Fix pause and resume
The BasicResults class was not calling Pause or Resume on
m_taskController. m_taskController is the instance of ITaskReader that
is passed to all of the worker classes that use it to check if they
should pause or stop.

Changed the BackendUploader to check TransferProgressAsync instead of
ProgressAsync, as the transfer function is meant for the worker that is
uploading files.

Added two new phases, Paused and Paused_WaitForUpload. These are used
to inform the user about the state of the operation when a pause has been requested.
The current transfer rate will display while waiting for uploads to finish.

The BackendUploader will now periodically timeout waiting for input (files ready
to upload) and will check if a pause is requested. It will then wait for all
current uploads to finish and signal to the IOperationProgressUpdater
that the phase should be changed to Paused.