Check for pause or stop while uploading
When all upload slots are full check for a pause or
stop request while waiting for an upload to complete.

After a Put call check if cancellation was requested before marking the
file as uploaded.

Handle a request to stop now by cancelling any uploads and then set the
flush request as complete. This prevents a possible deadlock by the
BackupHandler waiting for the flush request.