Catch TaskCanceledException when reading from channel
If a backup is canceled while reading from a channel, TaskCanceledException can be thrown.
Catch this exception so it does not bubble up to the user. If uploads have been canceled
it does not matter that the read was canceled. If uploads were not canceled, the read will
be tried again.