Exclude invalid Windows paths
This change explicity excludes paths ending in a dot or a space from
being considered for back up in Windows and generates a clear
warning to the user.

Without this change, under Windows, directories ending in a dot or a
space can trigger a constraint violation in Duplicati when updating
the SQLite database and in that case will abort the entire backup
(see, for example, duplicati/duplicati#3518 and
duplicati/duplicati#4229).  Also, files ending in a dot or a space can
silently fail to be backed up at all.

In Windows, files and directories ending in a dot (".") or a space (" ")
are problematic.

For example,
https://docs.microsoft.com/en-gb/windows/win32/fileio/naming-a-file
says: "Do not end a file or directory name with a space or a
period. Although the underlying file system may support such names,
the Windows shell and user interface does not."

Also, some .NET path manipulation functions do not support trailing
dots and spaces.  For example, under Windows,
Path.GetFullPath(@"C:\dir \file.") will return @"C:\dir\file"; i.e.,
the trailing dot and space have been stripped.

Though they are problematic, directories and files ending in a dot or
a space can be created on Windows in a number of ways; e.g., by
various programs (like Firefox), by using the Windows command line,
and by using environments like Cygwin, and so can trip up Duplicati if
they are not explicitly excluded.