Support Windows source paths starting with "\\?\"
Under Windows, when a backup path has been specified that uses the
"`\\?\`" prefix, change Duplicati to consistently record all files and
folders with that prefix.

Add option `--list-filter-literal` to the Duplicati.CommandLine.exe
"find" command to disable the check for regex, group, and wildcard
characters in the "`<filename>`" argument.

Change the GUI so that it does not mistake the question mark in the
"`\\?\`" prefix with a wildcard by changing the restore picker to pass
`--list-filter-literal=true` when obtaining folder contents.