Fix issue where remote URL defined using run-script-before was ignored.
Previously, we were declaring that the m_remoteurl field be passed by
reference to the Execute method.  However, the OnStart method declared
that the remoteurl variable was to be passed by reference.  As a result,
the caller (Controller.SetupCommonOptions) method never received the
updated value.

Now, the variables being passed by reference are made consistent.

This fixes #3651.