Always exclude job-specific SQLite database files from backup.
There is no benefit to backing these up.  If the database files need to
be restored, they should be recreated from the backend files to ensure
consistency.  When these are selected for backup, the operation often
hangs and never finishes.

We cannot use filters to unconditionally exclude files, as there exists
code paths where filters containing both includes and excludes have a
tendency to include unmatched paths, which is probably the safer thing
to do if a path is not matched by any (include or exclude) filter.

This fixes #1797 and #4308.