Update SSH.NET to version 2020.0.1.
From the release notes:

Even though RFC 4253 requires that an identification string MUST be
terminated by a carriage return and line feed, this fix restores support
for banners and identification strings that are only terminated by a
line feed.

This is a workaround for an issue in version 7.4 of OpenSSH which was
fixed in version 7.5.