Remove unit test result reporter job.
Although this displays the test failures more clearly, it doesn't handle
matrix builds that well.  The failures aren't really differentiated per
job (Linux vs. Windows), and this also adds a test result reporter job
(successful), as well as a Tests Results job, which fails if a test
fails.  This makes the list of jobs that appear next to a commit a bit
messy.