Enhanced error reporting for all Google Services by unpacking and displaying the HTTP response that Google Drive and Google API returns. These changes have been made across all Google web methods.
In addition to this I have looked further into the 403 (forbidden) error being returned from Google Drive. In my case the error being returned was "This file has been identified as malware or spam and cannot be downloaded". It appears that Google Drive checks files before allowing them to be downloaded. In order to bypass this error Google Drive has a query string parameter called "acknowledgeAbuse" that must be set if a file is detected as having spam/malware (the flag can't be set generally, it can only be set for files detected as having spam/malware). I have modified the code to add this flag if this error is detected but only if a new Google Drive Duplicati option called "acknowledge-abuse" is also enabled.

Finally I have replaced (automatically) all "var" statements with their native types. The "var" statements were making the code difficult to read particularly from GitHub because it created uncertainty around variable types, removing the use of "var" makes the code let ambiguous.