Change ngSelected to a dictionary, add pointers to parent for child nodes
The functionality should be the exact same after these changes.
Rather than building a dictionary object from the ngSelected array all the time, use a dictionary from the beginning.
The dictionary will be updated when the check state of a node changes, which will then trigger the partial map to be updated. The partial map only needs to be updated when the dictionary changes rather than also building it all the time.
Made a general cleanup pass of the file.
Remove unused and newly unneeded functions, renamed some variables.