Fix nodes under root not being deselected properly
If there are two or more nodes under the root node, select all of the nodes then deselect the root node, one or more of the nodes under the root node remains selected.
Fixed by removing all selected nodes under the root from the selected dictionary when selecting a node. The check is propogated all the way up to the root node and if all of its children are selected, they can be removed from the dictionary.

Also, prevented the full path of a file from being placed in the partial map as there is nothing beneath a file.
updateNodesFromMap will now only update the nodes (and their children) that are passed in. Or if none are passed in then the entire tree is updated.