Change file picker to handle versions properly
Versions are handled specially. If selected they are always put in ngSelected. Exception to this is if the default version, the latest version, is the only version selected.
In that case the file itself is put in ngSelected because that version of the file will be restored from the chosen backup.

Added logic changes for the above plus edge cases of e.g.
1. Check a file, the first version is automatically checked and the file is added to ngSelected.
2. Check the first version, the file is automatically checked and added to ngSelected.
3. Check the second version, that version is added to ngSelected and the file is checked.
4. The file already has been selected, meaning the default version is selected, and then another version is checked. The file is removed from ngSelected, the two versions are added to ngSelected.
5. Propagation logic when deselecting versions or files.