Restore selected versions of a file
The modified time of the file is appended to the restored filename when multiple versions of a file are being restored
If the version of a file that is selected to be restored is the version of the file that is in the chosen backup, then no special logic needs to be performed. It is a normal restore.
If the selected version is not in the chosen backup, the fileId is passed to the backend and used to restore that specific version of the file.

When any version is selected that is not the default version (the version in the chosen backup), then ALL selected versions for that file have the fileId sent to the backend. Additionally, if
more than one version is selected, and one of those is the default version, all fileIds are sent to the backend and the file from the chosen backup is "ignored".
The files from the backup are put into a temp table and since specific versions are going to be restored, the file with versions selected must be removed from that table so the specific versions of
the file can be put in the table and have the modified time be part of the filenames.