Add utility functions for unpadded ("raw") Base64 encoding
Base64 encoding (in the url safe dialect or standard) normally uses padding (up to two '=')
as necessary, but if the data length is known this can be omitted according to the standard.
The .NET converter ToBase64String includes padding (when necessary), and FromBase64String
throws "Invalid length" if padding is missing. But JWT tokens use Base64 url encoding
without padding so to be able to use standard converters padding must be explicitely handled.

This commit completes existing utility functions for base64 with variants that handle
unpadded variants. These are denoted "raw", a term borrowed from Golang's base64 package
which calls them "RawURLEncoding" and "RawStdEncoding".