Jottacloud backend oauth token access lock
Possible workaround, or temporary proof of concept, for issue with Jottacloud oauth
token handling.

The exact sequence is not known, but seems to be related to long-running backups, and
some 12 hours elapsed clock time. Authentication with Jottacloud will for some reason
start to fail and manually recreating cli token and authid is necessary.

Duplicati gets HTTP 500 (Internal Server Error) from the oauth service while it attempts
at refreshing the token, and the oauth service logs indicate it gets a
HTTP 400 (Bad Request) from Jottacloud with description:
{"error":"invalid_grant","error_description":"Stale token"}.

The theory is that Jottacloud, in contrast to most others, invalidates a re-used
refresh token - forcing it to expire in perhaps 12 hours. The theory further is that
multi-threading in Duplicati and/or the OAuth service leads to token being refreshed
by several concurrent requests, leading to some of them issuing a refresh request
containing an old refresh token.