Add unit tests for restoring specific versions of files, fix existing unit tests
Fixed existing unit tests by:
m_filters was null because the unit tests did not call a constructor on FilterExpression which set it. Create a new List on construction so it is not null.
If m_filters was empty, OnlyContainsFilterType() would return true. Added a check that there were items in the list first.
A call to Parameters.Clear() was accidentally removed after ExecuteScalarInt64() was called. ExecuteScalarInt64() clears parameters, adds any if needed, but does not clear its own parameters.
Use the proper OS directory separator before the &fileid= indicator from the frontend of what file id to restore.