Show files excluded via 'Exclude' section in source picker.
Add more metadata to the file list returned by server, so that the source file picker can display filters based on file attributes.
Correctly display 'X' in file list if a file is excluded by attributes and not included as a source (#4751).
Fix directory separator issue when loading the page and prevent opening folders in source file list.