Upgrading AWS SDK to latest
This also resolves a series of warnings produced by unequalized versions between AWSSDK.Core and AWSSDK.S3 references.

The decision to not upgrade Minio (which is rather dated) is because starting at 4.0.0 several calls were made obsolete including the constructor. A specific feature branch will be opened to integrate the new version and run CI tests.