Add feature to send messages to Log/Report from script
This works by parsing the stdout stream, line by line, looking for lines that start with the prefixes LOG:WARN, LOG:ERROR, LOG:INFO and forwarding them to the Log/Report.

In a script that could be used with a simple echo "LOG:WARN Some warning" or even if its an executable application, if it outputs to sdtout lines with the prefix, it will of course also be captured.