Removing timeout from Webdav
The timeout will be controlled externally by the TimeoutObservingStream and or the cancellation token.

Timeout for List set to default of 10 minutes

Since filesystems may take quite sometime to list large number of files, the timeout was set to a high value.

Previous timeout of 30 seconds was triggering in a case reported by user.

Timeout for CreateFolder and DeleteAsync was kept to default of 30 seconds.

Minor formating/refactoring was done.