Disabling unsupported platforms for CI
Currently the CIFS Tests on macos and windows require it to run in cloud mode, which means the port binding would not match default netbios/directtcp protocol standards as the container port allocation is random and the backend does not support specifying explicit ports.

The remaining tests on linux nevertheless already grants some degree of assurance the backend is working.